Turning the values back to ints due to how PPM works.

+++ Source/Camera/Camera.h
@@ -7,7 +7,7 @@ namespace RayTracer {
 
     class Camera {
     public:
-        Camera(float imageWidth, float imageHeight);
+        Camera(int imageWidth, int imageHeight);
 
         [[nodiscard]] static float FocalLength() ;
         [[nodiscard]] Point3 Center() const;
